refactor: package 이름 변경  - configure -> config
